requirements.txt
altgraph==0.16.1
astroid==2.3.1
attrs==19.3.0
auto-py-to-exe==2.6.5
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.3
bottle==0.12.17
bottle-websocket==0.2.9
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
Eel==0.10.4
entrypoints==0.3
Flask==1.1.2
flask-marshmallow==0.12.0
future==0.17.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.8
imageio==2.5.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.3
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
lazy-object-proxy==1.4.2
macholib==1.11
MarkupSafe==1.1.1
marshmallow==3.5.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.3
notebook==6.0.3
numpy==1.16.4
opencv-contrib-python==4.1.0.25
opencv-python==4.1.1+contrib
pandocfilters==1.4.2
parso==0.6.2
pefile==2019.4.18
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
pycparser==2.19
Pygments==2.6.1
PyInstaller==3.5
pylint==2.4.2
PyMySQL==0.9.3
PyQt5==5.13.0
PyQt5-sip==4.19.18
pyqt5-tools==5.11.3.1.4
PyQtWebEngine==5.13.1
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.10.3
pytube3==9.6.4
PyWavelets==1.0.3
pywin32==224
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzbar==0.1.8
pyzmq==19.0.0
requests==2.22.0
scikit-image==0.15.0
selenium==3.141.0
Send2Trash==1.5.0
shiboken2==5.13.0
six==1.12.0
soupsieve==1.9.3
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.0
typing-extensions==3.7.4.3
urllib3==1.25.3
virtualenv==16.7.9
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
whichcraft==0.6.0
wrapt==1.11.2
zipp==3.1.0
